PR2PO application added
It is a common practice in manufacturing companies to create purchase requisitions for the goods or
items required. A purchase requisition typically contains the item code of the good required, its
quantity and the date by when it is required. Based on the submitted purchase requisition, the Buyer
or the Purchasing Officer then generates a purchase order. A purchase order or PO is a legal docu-
ment used by organizations for placing orders with their suppliers for procurement of goods.

PR2PO is a command-driven procurement application that helps the users to
create a purchase requisition
view requisition details
convert purchase requisition into a purchase order